+++ src/components/common/Sections/HideContentButton.js
@@ -1,0 +1,31 @@
+import React from 'react'
+import styled from '@emotion/styled'
+import { UpOutlined, DownOutlined } from '@ant-design/icons'
+import { Button } from 'antd'
+
+const HideContentButton = styled(
+  ({ toggleContentVisibility, isContentVisible, ...props }) => (
+    <Button
+      {...props}
+      type="link"
+      icon={isContentVisible ? <UpOutlined /> : <DownOutlined />}
+      onClick={toggleContentVisibility}
+    />
+  )
+)`
+  float: right;
+  position: absolute;
+  top: 11px;
+  right: 12px;
+  font-size: 12px;
+  border-width: 0px;
+  width: 20px;
+  height: 20px;
+  color: rgba(0, 0, 0, 0.45);
+  &:hover,
+  &:focus {
+    color: #24292e;
+  }
+`
+
+export default HideContentButton
